django_app/requirements.txt
aiofiles==24.1.0
aioflask==0.4.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
click==8.1.8
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
Django==5.1.7
Flask==3.1.0
frozenlist==1.5.0
greenlet==3.1.1
greenletio==0.11.0
h11==0.14.0
h2==4.1.0
hpack==4.1.0
Hypercorn==0.17.3
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.1.0
nats-py==2.9.0
pillow==11.1.0
priority==2.0.0
propcache==0.2.1
psycopg[binary,pool]>=3.1.18
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyOpenSSL==25.0.0
service-identity==24.2.0
setuptools==76.1.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
uvicorn==0.34.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
zope.interface==7.2
channels>=4.0.0
channels-redis>=4.2.0
